redis wrapper working

+++ packages/server/src/utils/RedisContext.ts
@@ -1,48 +1,66 @@
-import { createClient, RedisClientType } from "redis";
+import { createClient, RedisClientType, SetOptions } from "redis";
 import { randomBytes } from "crypto";
 
-export default class RedisContext {
+export class RedisContext {
   keyPrefix: string;
   redisClient: RedisClientType;
-  isOpen = false;
-  static async build() {
+
+  static build(withPrefix?: boolean) {
     const keyPrefix = randomBytes(4).toString("hex");
-    return new RedisContext(keyPrefix);
-  }
-  constructor(keyPrefix: string) {
-    this.keyPrefix = keyPrefix;
-  }
-  connect() {
-    this.redisClient ==
+    return new RedisContext(
       createClient({
         url: process.env.REDIS_URL,
-      });
-    this.isOpen = true;
+      }),
+      withPrefix ? keyPrefix : ""
+    );
   }
-  disconnect() {
-    this.redisClient.disconnect();
-    this.isOpen = false;
+
+  constructor(redisClient: RedisClientType, keyPrefix: string) {
+    this.redisClient = redisClient;
+    this.keyPrefix = keyPrefix;
+    this.redisClient.on("error", (err) => console.log(err));
   }
-  set(key: string, value: any) {
-    this.redisClient.set(this.keyPrefix + key, value);
+
+  async connect() {
+    await this.redisClient.connect();
+    console.log(`redis client with ${this.keyPrefix ? `context ${this.keyPrefix}` : "vanilla context"} connected`);
   }
-  get(key: string) {
-    this.redisClient.get(this.keyPrefix + key);
+  async disconnect() {
+    await this.redisClient.disconnect();
+    console.log(`redis client with ${this.keyPrefix ? `context ${this.keyPrefix}` : "vanilla context"} disconnected`);
   }
-  del(key: string) {
-    this.redisClient.del(this.keyPrefix + key);
+  async set(key: string, value: any, options?: SetOptions | undefined) {
+    await this.redisClient.set(this.keyPrefix + key, value, options);
   }
-  async getByPattern() {
-    let currentCursor;
+  async get(key: string) {
+    return await this.redisClient.get(this.keyPrefix + key);
+  }
+  async del(key: string) {
+    await this.redisClient.del(this.keyPrefix + key);
+  }
+  async getKeysByPrefix() {
+    let currentCursor: number | undefined;
     let keysToReturn: string[] = [];
     while (currentCursor !== 0) {
-      const { cursor, keys } = await this.redisClient.scan(0, { MATCH: this.keyPrefix });
+      const { cursor, keys } = await this.redisClient.scan(currentCursor || 0, { MATCH: this.keyPrefix + "*", COUNT: 10 });
       keysToReturn = [...keysToReturn, ...keys];
       currentCursor = cursor;
     }
     return keysToReturn;
   }
   async unlink(keys: string[]) {
-    return await this.redisClient.unlink(keys);
+    if (keys.length) return await this.redisClient.unlink(keys);
+  }
+  async removeAllKeys() {
+    const keysToRemove = await this.getKeysByPrefix();
+    const numKeysRemoved = await this.unlink(keysToRemove);
+  }
+  async cleanup() {
+    await this.removeAllKeys();
+    await this.disconnect();
   }
 }
+
+export const wrappedRedis: { [client: string]: RedisContext | undefined } = {
+  context: undefined,
+};
